Refactoring and added the alert functionality to reportScreen
Removed the alert implementation entirely from AppNavigation.js, so that the implementation is in one place only (functions/handleBack.js, that is used in the component at components/ReportEditingBackButton.js). The same implementation is also used in ReportScreen, but isUnsaved is temporarily hardcoded to false so that the alert is not shown.