Unsaved-alert (#68)
* Fixed IOS OfflineNotice bug (switches on/off in every screen), added clarification for the alert box that appears when user tries to leave ReportScreen without saving.

* TODO messages and clarifications on the bugs found

* IOS bug fix - could not press on the Icon in the Templates header

* Added default value for handleBack and added the new alert text

* Moved handleBack to a separate functions folder

* Renamed headerBackButtonWrapper and moved it to a dedicated compoment folder.

The new name is ReportEditingBackButton. This is exported on one line with redux's connect, so
now the name HeaderBackButtonWrapperWithRedux isn't used anywhere either."

* Refactoring and added the alert functionality to reportScreen

Removed the alert implementation entirely from AppNavigation.js, so that the implementation is in one place only (functions/handleBack.js, that is used in the component at components/ReportEditingBackButton.js). The same implementation is also used in ReportScreen, but isUnsaved is temporarily hardcoded to false so that the alert is not shown.

* Moved saving functionality to the back handling function (handleBack)

Now the saving functionality works from the alert. There are still some commented methods in saving
that don't work, but they might be redundant. Currently saving is also implemented in some screen
files, and I think it would be nice to unify the implementations.

* Added some comments

* Possible saving solution through checking redux flag on unmount

* Fixed bug in report saving

* Moved some variables out of newReport reducer

isUnsaved and isSavingRequested are now moved to their own reducer. This is so that these values don't get saved inside the reports along with actual report data

* Some cleaning up

* Fixed bug in saving and editing reports

* Prevent empty drafts from being saved in saveDraft (api.js)